new preset manager + correct edit param handling
