+++ for Arduino/photopizza_v1_2_2/photopizza_v1_2_2.ino
@@ -20,6 +20,7 @@
 void setup() {
   Serial.begin(115200);
 
+  //delay(1000); //for debug
   libInit();
   IrInit();
 }
