+++ for Arduino/photopizza_v1_2_2/presetManager.h
@@ -4,6 +4,12 @@
 #include "defines.h"
 #include "eepromAnything.h"
 
+#include "param.h"
+#include "paramAcc.h"
+#include "paramSteps.h"
+#include "paramSpeed.h"
+#include "paramDir.h"
+
 namespace PhotoPizza {
 
 typedef enum {
@@ -14,210 +20,17 @@ typedef enum {
   PARAM_COUNT,
 } paramType;
 
-struct presetStorage {
-//public:
+struct presetStorageData {
   long _speed; // speed
   long _steps; // rotaion
   long _acc; // acceleration
-  int _dir; // -1 - clockwise , 1 - counterclockwise
 };
 
-class param {
+struct presetStorage {
+  byte flag;
+  int  version;
 
-public:
-
-  param(){
-    _valStep = 1;
-    _val = 0;
-    _tmpVal = 0;
-    _valHiLimit = 100;
-    _valLoLimit = 0;
-    _edit = false;
-  }
-
-  virtual void up(){
-    long tmp = get();
-       tmp = tmp - tmp % _valStep + _valStep;
-    if(tmp > _valHiLimit)
-      tmp = _valLoLimit;
-    set(tmp);
-  }
-  virtual void down(){
-    long tmp = get();
-      tmp = tmp - tmp % _valStep - _valStep;
-    if(tmp < _valLoLimit)
-      tmp = _valHiLimit;
-    set(tmp);
-  }
-
-  virtual void edit(){
-    _tmpVal = _val;
-    _edit = true;
-  }
-
-  virtual bool save(){
-    _edit = false;
-    return set(_tmpVal);
-  }
-
-  virtual void discard(){
-    _edit = false;
-  }
-
-  virtual String ToString(bool shorten = false){
-    if(_edit)
-      return (String) _tmpVal;
-    else
-      return (String) _val;
-  }
-  virtual String getName(bool shorten = false){
-    return "param";
-  }
-
-  virtual bool set(long val){
-    if (val < _valLoLimit)
-      val = _valLoLimit;
-    if (val > _valHiLimit)
-      val = _valHiLimit;
-    if(_edit)
-      _tmpVal = val;
-    else
-      _val = val;
-    return true;
-  }
-  virtual long get() {
-    if(_edit)
-      return _tmpVal;
-    else
-      return _val;
-  }
-
-  virtual operator long() {
-    return get();
-  }
-
-  virtual param& operator=( long val ) {
-    set(val);
-    return *this;
-  }
-
-  virtual bool isEdit(){
-    return _edit;
-  }
-
-  virtual ~param(){};
-
-protected:
-  long _valHiLimit;
-  long _valLoLimit;
-  long _valStep;
-  long _val;
-  bool _edit;
-  long _tmpVal;
-};
-
-class paramSpeed : public param {
-public:
-  paramSpeed() : paramSpeed(1000){}
-  paramSpeed(long val){
-    _valStep = SPEED_STEP;
-    _valLoLimit = SPEED_MIN;
-    _valHiLimit = SPEED_MAX;
-    this->set(val);
-  }
-
-  virtual String getName(bool shorten = false){
-    return "speed";
-  }
-};
-
-class paramSteps : public param {
-public:
-  paramSteps() : paramSteps(3200){}
-  paramSteps(long val){
-    _valStep = ROT_STEP;
-    _valLoLimit = ROT_MIN;
-    _valHiLimit = ROT_MAX;
-    this->set(val);
-  }
-
-  virtual String getName(bool shorten = false){
-    return "steps";
-  }
-
-  virtual String ToString(bool shorten = false){
-    long val = get();
-    if(val == 0)
-      return "inf";
-    else
-      return (String) val;
-  }
-};
-
-class paramAcc : public param {
-public:
-  paramAcc() : paramAcc(5000){}
-  paramAcc(long val){
-    _valStep = ACC_STEP;
-    _valLoLimit = ACC_MIN;
-    _valHiLimit = ACC_MAX;
-    this->set(val);
-  }
-
-  virtual String getName(bool shorten = false){
-    return "accel";
-  }
-
-  virtual String ToString(bool shorten = false){
-    long val = get();
-    if(val > 0)
-      return (String)val;
-    else if(val == 0)
-      return "inf";
-    else
-      return "?";
-  }
-};
-
-#define PARAM_DIR_VAL_COUNT 2
-
-typedef struct enumParamMap{
-  long   value;
-  char*  label;
-};
-
-class paramDir : public param {
-public:
-  paramDir() : paramDir(CW){}
-  paramDir(long val): _map{
-    {CW, "CW"}, {CCW, "CCW"}
-  }{
-    _valLoLimit = 0;
-    _valHiLimit = PARAM_DIR_VAL_COUNT - 1;
-    _valStep = 1;
-    this->set(val);
-  }
-
-  virtual String getName(bool shorten = false){
-    return "dir";
-  }
-
-  virtual long get(){
-    if(_edit)
-      return _map[_tmpVal].value;
-    else
-      return _map[_val].value;
-  }
-
-  virtual String ToString(bool shorten = false){
-    if(_edit)
-       return _map[_tmpVal].label;
-    else
-      return _map[_val].label;
-  }
-
-protected:
-  enumParamMap _map[PARAM_DIR_VAL_COUNT];
+  presetStorageData data[NUM_PROGRAMS];
 };
 
 class preset {
@@ -240,6 +53,52 @@ public:
     default:
       return _default;
     }
+  }
+
+  preset& operator=(int val[4]) {
+    //Serial.println("op = ar");
+    _speed = val[0];
+    _acc = val[1];
+    _steps = val[2];
+    _dir = val[3];
+    return *this;
+  }
+
+  bool operator!=(presetStorageData & val) {
+    if(_speed != val._speed)
+       return true;
+    if(_acc != val._acc)
+       return true;
+    if(_steps * _dir != val._steps)
+       return true;
+
+    return false;
+  }
+
+  operator presetStorageData() {
+    //Serial.println("op = cast to psd");
+    presetStorageData tmp;
+    tmp._acc = _acc;
+    tmp._speed = _speed;
+    tmp._steps = _steps * _dir;
+    return tmp;
+  }
+
+  preset& operator=(presetStorageData & val) {
+    /*Serial.println("op = psd");
+    Serial.println((String)"Sp: " + val._speed);
+    Serial.println((String)"acc: " + val._acc);
+    Serial.println((String)"steps: " + val._steps);*/
+    _speed = val._speed;
+    _acc = val._acc;
+
+    if(val._steps < 0)
+      _dir = CW;
+    else
+      _dir = CCW;
+
+    _steps = abs(val._steps);
+    return *this;
   }
 
 private:
@@ -290,7 +149,7 @@ public:
 
   void changeDirection(int dir);
 
-  void save();
+  void save(bool force = false);
 
 private:
   int _curPreset; // current preset
